revert subcomponent changes
