+++ frontend/src/components/index.js
@@ -1,10 +1,12 @@
 import Dashboard from './Dashboard'
+import DocumentList from './DocumentList'
+import DocumentListItem from './DocumentListItem'
 import DocumentPreview from './DocumentPreview'
 
 import SelectDocumentsPage from './SelectDocuments'
 import BoxPreview from './DocumentPreview'
 import PMMainPage from './PMMainPage'
-import DocumentClassPage from './DocumentClass'
+import DocumentClassPage from './DocumentClassPage.js'
 
 import Load from './Load'
 
@@ -15,16 +17,18 @@ import Register from './auth/Register'
 import WrongPage from './auth/WrongPage'
 
 export {
-  ErrorMessage,
-  ForgotPassword,
-  LogIn,
-  Register,
-  WrongPage,
   Dashboard,
+  DocumentList,
+  DocumentListItem,
+  LogIn,
   BoxPreview,
+  Register,
   SelectDocumentsPage,
   PMMainPage,
+  ForgotPassword,
   DocumentClassPage,
   DocumentPreview,
-  Load
+  Load,
+  WrongPage,
+  ErrorMessage
 }
