+++ frontend/src/tests/DocumentList.test.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { DocumentList } from '../components/Dashboard'
+import { DocumentList } from '../components/DocumentList'
 import Adapter from 'enzyme-adapter-react-16'
 import { shallowToJson } from 'enzyme-to-json'
 import { configure, shallow } from 'enzyme'
